Cleaned up Repository Directory and File Structure
