+++ CPP/Algorithms/Others/lexicographical_compare.cpp
