+++ CPP/Algorithms/Others/remove_copy_if.cpp
