+++ CPP/Algorithms/Others/replace_copy_if.cpp
