+++ CPP/Algorithms/Searching/linearSearch.cpp
