+++ JAVA/Algorithms/Maths/PerfectSquares.java
