+++ JAVA/OOP/Composition.java
