+++ PYTHON/Algorithms/Arrays/AlignmentProblem-CreatingADIstanceMatrix.py
